rewrite openbsd_service provider using rcctl
In OpenBSD 5.7 a new utility `rcctl(8)` was introduced to provide a
a cleaner interface to managed base and package daemon control scripts.

The previous method for managing services on OpenBSD involved reading
and writing to the `/etc/rc.conf` and `/etc/rc.conf.local` files. Using
`rcctl(8)` simplifies the openbsd_service provider and hopefully
makes it a little more future proof.

Additionally, the previous openbsd_service provider did not support
setting `flags` on services (AFAICT). This is important on OpenBSD
for a service like dhcpd where you would typically set a flag such as:

    # /etc/rc.conf/local
    dhcpd_flags="em1"

This would be passed as an arg to dhcpd on start.

I implemented support for this using the existing `parameters` attribute, eg:

    service 'dhcpd' do
      parameters(flags: 'em1')
      action [:enable, :start]
    end

The only vars supported in OpenBSD 5.7 by `rcctl(8)` are: `flags`, `timeout`, `user`.
The provider does not specifically care about what keys are passed to the `parameters`
attribute at this time, for future compatibility.

Questions/TODOs:

- `rcctl(8)` is only available on OpenBSD 5.7+. Should this be an additional provider
  that is only used as the default if 5.7+ is detected? We could raname this to
  `Chef::Provider::Service::OpenbsdRcctl` to not conflict with the old version.
- Right now parameters are only checked and set on `action_enable` since `rcctl(8)`
  only allows setting parameters on enabled services.. It might be
  desirable to also set parameters in `action_start` in the case that `action_start`
  is processed first and causes the daemon start to fail. Same question for `action_restart`.
- What/where would be the best place to document the use of `parameters` attribute
  for this service? It appears somewhat unique amongst service providers from a
  quick survey.
- Not all rc.d scripts support `reload`. We might be able to detect this and remove
  it from the list of :supported actions. As written, if you try to reload a service
  that doesn't support it the underlying rc.d/rcctl calls will fail anyway.